+++ src/spyglass/common/common_filter.py
@@ -212,7 +212,7 @@ class FirFilterParameters(SpyglassMixin, dj.Manual):
             start = all[0]
 
         if stop > all[-1]:
-            warnings.warn(
+            logger.warning(
                 timestamp_warn
                 + "stop time larger than last timestamp, "
                 + f"substituting last: {stop} < {all[-1]}"