+++ src/spyglass/spikesorting/imported.py
@@ -1,5 +1,8 @@
+import copy
+
 import datajoint as dj
 import pynwb
+from datajoint.utils import to_camel_case
 
 from spyglass.common.common_nwbfile import Nwbfile
 from spyglass.common.common_session import Session  # noqa: F401
@@ -16,18 +19,28 @@ class ImportedSpikeSorting(SpyglassMixin, dj.Imported):
     object_id: varchar(40)
     """
 
+    _nwb_table = Nwbfile
+
     def make(self, key):
+        orig_key = copy.deepcopy(key)
         nwb_file_abs_path = Nwbfile.get_abs_path(key["nwb_file_name"])
 
         with pynwb.NWBHDF5IO(
             nwb_file_abs_path, "r", load_namespaces=True
         ) as io:
             nwbfile = io.read()
-            if nwbfile.units:
-                key["object_id"] = nwbfile.units.object_id
-                self.insert1(key, skip_duplicates=True)
-            else:
+            if not nwbfile.units:
                 logger.warn("No units found in NWB file")
+                return
+
+        from spyglass.spikesorting.merge import SpikeSortingOutput  # noqa: F401
+
+        key["object_id"] = nwbfile.units.object_id
+        part_name = SpikeSortingOutput._part_name(self.table_name)
+        self.insert1(key, skip_duplicates=True)
+        SpikeSortingOutput._merge_insert(
+            [orig_key], part_name=part_name, skip_duplicates=True
+        )
 
     @classmethod
     def get_recording(cls, key):