+++ src/spyglass/utils/dj_merge_tables.py
@@ -1,7 +1,9 @@
 import re
 from contextlib import nullcontext
+from inspect import getmodule
 from itertools import chain as iter_chain
 from pprint import pprint
+from typing import Union
 
 import datajoint as dj
 from datajoint.condition import make_condition
@@ -10,7 +12,6 @@ from datajoint.preview import repr_html
 from datajoint.utils import from_camel_case, get_master, to_camel_case
 from IPython.core.display import HTML
 
-from spyglass.utils.dj_helper_fn import fetch_nwb
 from spyglass.utils.logging import logger
 
 RESERVED_PRIMARY_KEY = "merge_id"
@@ -51,21 +52,45 @@ class Merge(dj.Manual):
                         + f"\n\tExpected: {self.primary_key}"
                         + f"\n\tActual  : {part.primary_key}"
                     )
-        self._analysis_nwbfile = None
-
-    @property  # CB: This is a property to avoid circular import
-    def analysis_nwbfile(self):
-        if self._analysis_nwbfile is None:
-            from spyglass.common import AnalysisNwbfile  # noqa F401
-
-            self._analysis_nwbfile = AnalysisNwbfile
-        return self._analysis_nwbfile
+        self._source_class_dict = {}
 
     def _remove_comments(self, definition):
         """Use regular expressions to remove comments and blank lines"""
         return re.sub(  # First remove comments, then blank lines
             r"\n\s*\n", "\n", re.sub(r"#.*\n", "\n", definition)
         )
+
+    @staticmethod
+    def _part_name(part=None):
+        """Return the CamelCase name of a part table"""
+        if not isinstance(part, str):
+            part = part.table_name
+        return to_camel_case(part.split("__")[-1].strip("`"))
+
+    def get_source_from_key(self, key: dict) -> str:
+        """Return the source of a given key"""
+        return self._normalize_source(key)
+
+    def parts(self, camel_case=False, *args, **kwargs) -> list:
+        """Return a list of part tables, add option for CamelCase names.
+
+        See DataJoint `parts` for additional arguments. If camel_case is True,
+        forces return of strings rather than objects.
+        """
+        self._ensure_dependencies_loaded()
+
+        if camel_case and kwargs.get("as_objects"):
+            logger.warning(
+                "Overriding as_objects=True to return CamelCase part names."
+            )
+            kwargs["as_objects"] = False
+
+        parts = super().parts(*args, **kwargs)
+
+        if camel_case:
+            parts = [self._part_name(part) for part in parts]
+
+        return parts
 
     @classmethod
     def _merge_restrict_parts(
@@ -220,33 +245,32 @@ class Merge(dj.Manual):
             for p in cls._merge_restrict_parts(
                 restriction=restriction,
                 add_invalid_restrict=False,
-                return_empties=True,
+                return_empties=False,  # motivated by SpikeSortingOutput.Import
             )
         ]
 
-        primary_attrs = list(
-            dict.fromkeys(  # get all columns from parts
-                iter_chain.from_iterable([p.heading.names for p in parts])
+        attr_dict = {  # NULL for non-numeric, 0 for numeric
+            attr.name: "0" if attr.numeric else "NULL"
+            for attr in iter_chain.from_iterable(
+                part.heading.attributes.values() for part in parts
             )
-        )
-        # primary_attrs.append(cls()._reserved_sk)
-        query = dj.U(*primary_attrs) * parts[0].proj(  # declare query
-            ...,  # include all attributes from part 0
-            **{
-                a: "NULL"  # add null value where part has no column
-                for a in primary_attrs
-                if a not in parts[0].heading.names
-            },
-        )
-        for part in parts[1:]:  # add to declared query for each part
-            query += dj.U(*primary_attrs) * part.proj(
-                ...,
+        }
+
+        def _proj_part(part):
+            """Project part, adding NULL/0 for missing attributes"""
+            return dj.U(*attr_dict.keys()) * part.proj(
+                ...,  # include all attributes from part
                 **{
-                    a: "NULL"
-                    for a in primary_attrs
-                    if a not in part.heading.names
+                    k: v
+                    for k, v in attr_dict.items()
+                    if k not in part.heading.names
                 },
             )
+
+        query = _proj_part(parts[0])  # start with first part
+        for part in parts[1:]:  # add remaining parts
+            query += _proj_part(part)
+
         return query
 
     @classmethod
@@ -294,7 +318,7 @@ class Merge(dj.Manual):
             keys = []  # empty to-be-inserted key
             for part in parts:  # check each part
                 part_parent = part.parents(as_objects=True)[-1]
-                part_name = to_camel_case(part.table_name.split("__")[-1])
+                part_name = cls._part_name(part)
                 if part_parent & row:  # if row is in part parent
                     if keys and mutual_exclusvity:  # if key from other part
                         raise ValueError(
@@ -475,16 +499,18 @@ class Merge(dj.Manual):
         for part_parent in part_parents:
             super().delete(part_parent, **kwargs)
 
-    @classmethod
     def fetch_nwb(
-        cls,
+        self,
         restriction: str = True,
         multi_source=False,
         disable_warning=False,
         *attrs,
         **kwargs,
     ):
-        """Return the AnalysisNwbfile file linked in the source.
+        """Return the (Analysis)Nwbfile file linked in the source.
+
+        Relies on SpyglassMixin._nwb_table_tuple to determine the table to
+        fetch from and the appropriate path attribute to return.
 
         Parameters
         ----------
@@ -493,32 +519,14 @@ class Merge(dj.Manual):
         multi_source: bool
             Return from multiple parents. Default False.
         """
-        if not disable_warning:
-            _warn_on_restriction(table=cls, restriction=restriction)
+        if isinstance(self, dict):
+            raise ValueError("Try replacing Merge.method with Merge().method")
+        if restriction is True and self.restriction:
+            if not disable_warning:
+                _warn_on_restriction(self, restriction)
+            restriction = self.restriction
 
-        part_parents = cls._merge_restrict_parents(
-            restriction=restriction,
-            return_empties=False,
-            add_invalid_restrict=False,
-        )
-
-        if not multi_source and len(part_parents) != 1:
-            raise ValueError(
-                f"{len(part_parents)} possible sources found in Merge Table:"
-                + " and ".join([p.full_table_name for p in part_parents])
-            )
-
-        nwbs = []
-        for part_parent in part_parents:
-            nwbs.extend(
-                fetch_nwb(
-                    part_parent,
-                    (cls().analysis_nwbfile, "analysis_file_abs_path"),
-                    *attrs,
-                    **kwargs,
-                )
-            )
-        return nwbs
+        return self.merge_restrict_class(restriction).fetch_nwb()
 
     @classmethod
     def merge_get_part(
@@ -527,6 +535,7 @@ class Merge(dj.Manual):
         join_master: bool = False,
         restrict_part=True,
         multi_source=False,
+        return_empties=False,
     ) -> dj.Table:
         """Retrieve part table from a restricted Merge table.
 
@@ -545,6 +554,8 @@ class Merge(dj.Manual):
             native part table.
         multi_source: bool
             Return multiple parts. Default False.
+        return_empties: bool
+            Default False. Return empty part tables.
 
         Returns
         ------
@@ -563,11 +574,11 @@ class Merge(dj.Manual):
             restricting
         """
         sources = [
-            to_camel_case(n.split("__")[-1].strip("`"))  # friendly part name
-            for n in cls._merge_restrict_parts(
+            cls._part_name(part)  # friendly part name
+            for part in cls._merge_restrict_parts(
                 restriction=restriction,
                 as_objects=False,
-                return_empties=False,
+                return_empties=return_empties,
                 add_invalid_restrict=False,
             )
         ]
@@ -595,7 +606,9 @@ class Merge(dj.Manual):
         cls,
         restriction: str = True,
         join_master: bool = False,
-        multi_source=False,
+        multi_source: bool = False,
+        return_empties: bool = False,
+        add_invalid_restrict: bool = True,
     ) -> dj.FreeTable:
         """Returns a list of part parents with restrictions applied.
 
@@ -610,6 +623,12 @@ class Merge(dj.Manual):
             Default True.
         join_master: bool
             Default False. Join part with Merge master to show uuid and source
+        multi_source: bool
+            Return multiple parents. Default False.
+        return_empties: bool
+            Default False. Return empty parent tables.
+        add_invalid_restrict: bool
+            Default True. Include parent for which the restriction is invalid.
 
         Returns
         ------
@@ -620,11 +639,12 @@ class Merge(dj.Manual):
         part_parents = cls._merge_restrict_parents(
             restriction=restriction,
             as_objects=True,
-            return_empties=False,
-            add_invalid_restrict=False,
+            return_empties=return_empties,
+            add_invalid_restrict=add_invalid_restrict,
         )
 
         if not multi_source and len(part_parents) != 1:
+            __import__("pdb").set_trace()
             raise ValueError(
                 f"Found  {len(part_parents)} potential parents: {part_parents}"
                 + "\n\tTry adding a string restriction when invoking "
@@ -636,6 +656,71 @@ class Merge(dj.Manual):
             part_parents = [cls * part for part in part_parents]
 
         return part_parents if multi_source else part_parents[0]
+
+    @property
+    def source_class_dict(self) -> dict:
+        if not self._source_class_dict:
+            module = getmodule(self)
+            self._source_class_dict = {
+                part_name: getattr(module, part_name)
+                for part_name in self.parts(camel_case=True)
+            }
+        return self._source_class_dict
+
+    def _normalize_source(
+        self, source: Union[str, dj.Table, dj.condition.AndList, dict]
+    ) -> str:
+        fetched_source = None
+        if isinstance(source, (Merge, dj.condition.AndList)):
+            try:
+                fetched_source = (self & source).fetch(self._reserved_sk)
+            except DataJointError:
+                raise ValueError(f"Unable to find source for {source}")
+            source = fetched_source[0]
+            if len(fetched_source) > 1:
+                logger.warn(f"Multiple sources. Selecting first: {source}.")
+        if isinstance(source, dj.Table):
+            source = self._part_name(source)
+        if isinstance(source, dict):
+            source = self._part_name(self.merge_get_parent(source))
+
+        return source
+
+    def merge_get_parent_class(self, source: str) -> dj.Table:
+        """Return the class of the parent table for a given CamelCase source.
+
+        Parameters
+        ----------
+        source: Union[str, dict, dj.Table]
+            Accepts a CamelCase name of the source, or key as a dict, or a part
+            table.
+
+        Returns
+        -------
+        dj.Table
+            Class instance of the parent table, including class methods.
+        """
+
+        ret = self.source_class_dict.get(self._normalize_source(source))
+        if not ret:
+            logger.error(
+                f"No source class found for {source}: \n\t"
+                + f"{self.parts(camel_case=True)}"
+            )
+        return ret
+
+    def merge_restrict_class(self, key: dict) -> dj.Table:
+        """Returns native parent class, restricted with key."""
+        parent_key = self.merge_get_parent(key).fetch("KEY", as_dict=True)
+
+        if len(parent_key) > 1:
+            raise ValueError(
+                f"Ambiguous entry. Data has mult rows in parent:\n\tData:{key}"
+                + f"\n\t{parent_key}"
+            )
+
+        parent_class = self.merge_get_parent_class(key)
+        return parent_class & parent_key
 
     @classmethod
     def merge_fetch(self, restriction: str = True, *attrs, **kwargs) -> list: