+++ tests/conftest.py
@@ -78,7 +78,7 @@ def pytest_configure(config):
     os.environ["SPYGLASS_BASE_DIR"] = str(BASE_DIR)
 
     SERVER = DockerMySQLManager(
-        restart=False,
+        restart=TEARDOWN,
         shutdown=TEARDOWN,
         null_server=config.option.no_server,
         verbose=VERBOSE,
@@ -162,10 +162,13 @@ def server(request, teardown):
 @pytest.fixture(scope="session")
 def dj_conn(request, server, verbose, teardown):
     """Fixture for datajoint connection."""
-    config_file = "dj_local_conf.json_pytest"
+    config_file = "dj_local_conf.json_test"
+    if Path(config_file).exists():
+        os.remove(config_file)
 
     dj.config.update(server.creds)
     dj.config["loglevel"] = "INFO" if verbose else "ERROR"
+    dj.config["custom"]["spyglass_dirs"] = {"base": str(BASE_DIR)}
     dj.config.save(config_file)
     dj.conn()
     yield dj.conn()
@@ -242,6 +245,7 @@ def mini_closed(mini_path):
 def mini_insert(mini_path, teardown, server, dj_conn):
     from spyglass.common import Nwbfile, Session  # noqa: E402
     from spyglass.data_import import insert_sessions  # noqa: E402
+    from spyglass.spikesorting.merge import SpikeSortingOutput  # noqa: E402
     from spyglass.utils.nwb_helper_fn import close_nwb_files  # noqa: E402
 
     dj_logger.info("Inserting test data.")